Day 08 - tough one!
Approach:
First determine for the 10 signals, which value they represent:

1. The 1, 4, 7 and 8 were straightforward: their string length only occurs once.
2. 9, 0 and 6 are all 6 digits:
  - only the 6 includes the entire 4
  - only the 0 includes the entire 7
  - the 6 is the last string with 6 digits
3. 2, 3 and 5 are all 4 digits:
  - only the 3 includes the entire 1
  - only 2 includes the part that is NOT part of 9
  - the 5 is the last string with 5 digits

Map all those 10 signals onto a zero-based list, but format the signals alphabetically for easier matching.

Go through the four output values. Sort each value alphabetically. Search for them in the signal list.
-> Thinking now: making the list a dictionary with the signal as key makes the lookup easier.

Having the digits for the four output values results in a number. Reducing all numbers to a total is the answer!